requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.7.0
    # via flask
bootstrap-flask==2.3.3
    # via -r requirements.in
click==8.1.7
    # via flask
flask==3.0.0
    # via
    #   -r requirements.in
    #   bootstrap-flask
    #   flask-login
    #   flask-moment
    #   flask-sqlalchemy
    #   flask-wtf
flask-login==0.6.3
    # via -r requirements.in
flask-moment==1.0.5
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
flask-wtf==1.2.1
    # via -r requirements.in
greenlet==3.0.2
    # via sqlalchemy
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
    #   wtforms
packaging==23.2
    # via flask-moment
pyjwt==2.8.0
    # via -r requirements.in
sqlalchemy==2.0.23
    # via flask-sqlalchemy
typing-extensions==4.9.0
    # via sqlalchemy
werkzeug==3.0.1
    # via
    #   flask
    #   flask-login
wtforms==3.1.1
    # via
    #   bootstrap-flask
    #   flask-wtf
